+++ packages/sitecore-jss-nextjs/src/editing/index.ts
@@ -15,6 +15,7 @@ export {
   editingDataService,
 } from './editing-data-service';
 export { VercelEditingDataCache } from './vercel-editing-data-cache';
+export { FEAASRenderMiddleware, FEAASRenderMiddlewareConfig } from './feaas-render-middleware';
 export {
   EditingConfigMiddleware,
   EditingConfigMiddlewareConfig,